requirements.txt
fpdf
Pillow
PyPDF2
docx
python-docx
requests
moviepy